% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval for the mean}
\usage{
myci(x, conf.level = 0.95)
}
\arguments{
\item{x}{A numeric vector that is the sample data}

\item{conf.level}{A numeric value between 0 and 1 specifying the confidence level. Default is 0.95}
}
\value{
vector with lower and upper bounds of the confidence interval
}
\description{
calculates a (1 - alpha) * 100% confidence interval for the mean
of a single sample
}
\examples{
set.seed(23);x = rnorm(30,mean=10,sd=12)
myci(x)
}
