% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the sample size}

\item{iter}{this is the number of iterations}

\item{lambda}{the rate/mean of the Poisson distribution}

\item{...}{}
}
\value{
a set of plots (a histogram, a bar plot, a probability function plot for Poisson distribution)
}
\description{
This function demonstrates the Central Limit Theorem. It gets random samples from a Poisson distribution,
calculates the sample means, and generates histograms of these means. It also compares the sample distribution
to a theoretical normal distribution and includes a bar plot of the sampled Poisson values.
}
\examples{
mycltp(n=5, iter =1000, lambda=5)
}
