% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{this is the number of iterations}

\item{N}{the total number in a population}

\item{r}{the number of successful items}

\item{n}{the sample size}
}
\value{
a barplot of the proportions
}
\description{
Hypergeometric function
}
\examples{
myhyper(iter=100,N=20,r=12,n=5)
}
