% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Maximum Likelihood}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{A likelihood function to be evaluated}

\item{theta}{numeric vector}
}
\value{
The `theta` value corresponding to the maximum likelihood
}
\description{
This function computes the maximum likelihood estimate for a given likelihood function over a range of theta values.
}
\examples{
logbin2=function(theta){log(dbinom(1,prob=theta,size=2)) + log(dbinom(7,prob=theta,size=8))}
mymaxlikg(theta=seq(0,1,length=10000), lfun=logbin2)
}
