% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My curve
instructions:	display the curve, shaded area between the curve and
x axis from -∞ to x=a, and calculate the area (probability, P(X<=a))
which is released to the command-line in a list.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{x-value up to which the area under the curve is shaded and the probability is calculated}
}
\value{
A list of mu, standard deviation, and probability
}
\description{
My curve
instructions:	display the curve, shaded area between the curve and
x axis from -∞ to x=a, and calculate the area (probability, P(X<=a))
which is released to the command-line in a list.
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1.5)
}
