% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Display P-value areas
the function calculates the p-value, displays the corresponding t-distribution, and gives the graph with the p-value areas.}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{Numeric. The observed t-statistic.}

\item{xmax}{Numeric. The range of x-axis values to display, it's set to 4.}

\item{n}{Integer. This is the sample size, set to 20.}

\item{alpha}{Numeric. Significance level for the test, set to 0.05.}
}
\value{
A list containing the critical value q and p-value along with the graph showing the area of the p-value.
}
\description{
Display P-value areas
the function calculates the p-value, displays the corresponding t-distribution, and gives the graph with the p-value areas.
}
\examples{
mypvalue(t0 = 2, xmax = 4, n = 30, alpha = 0.05)
}
