% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in flight}

\item{gamma}{is the probability a plane will be truly overbooked(people > seats)}

\item{p}{probability of a "show"}
}
\value{
A list containing the optimal number of tickets
}
\description{
Ntickets
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
